refactor: 어댑터 패키지 분리
