+++ member_service/src/main/java/com/luckytree/member_service/member/adapter/data/MemberResponse.java
@@ -1,6 +1,6 @@
 package com.luckytree.member_service.member.adapter.data;
 
-import com.luckytree.member_service.member.adapter.persistence.MemberEntity;
+import com.luckytree.member_service.member.adapter.jpa.MemberEntity;
 import com.luckytree.member_service.member.domain.Member;
 import lombok.*;
