+++ member_service/src/main/java/com/luckytree/member_service/member/adapter/jpa/BookmarkEntity.java
@@ -1,4 +1,4 @@
-package com.luckytree.member_service.member.adapter.persistence;
+package com.luckytree.member_service.member.adapter.jpa;
 
 import com.luckytree.member_service.common.enums.Category;
 import com.luckytree.member_service.member.adapter.data.CreateBookmarkRequest;
