+++ member_service/src/main/java/com/luckytree/member_service/member/adapter/jpa/MemberEntity.java
@@ -1,4 +1,4 @@
-package com.luckytree.member_service.member.adapter.persistence;
+package com.luckytree.member_service.member.adapter.jpa;
 
 import com.luckytree.member_service.member.adapter.data.SignupRequest;
 import com.luckytree.member_service.member.domain.Member;
