--- member_service/src/main/java/com/luckytree/member_service/member/adapter/persistence/BookmarkIdInterface.java
@@ -1,5 +1,0 @@
-package com.luckytree.member_service.member.adapter.persistence;
-
-public interface BookmarkIdInterface {
-    long shop_id();
-}
