+++ member_service/src/main/java/com/luckytree/member_service/member/application/port/outgoing/MemberPort.java
@@ -1,11 +1,6 @@
 package com.luckytree.member_service.member.application.port.outgoing;
 
-import com.luckytree.member_service.member.adapter.persistence.BookmarkEntity;
-import com.luckytree.member_service.member.adapter.persistence.MemberEntity;
-import com.luckytree.member_service.member.domain.Bookmark;
-import com.luckytree.member_service.member.domain.Member;
-
-import java.util.List;
+import com.luckytree.member_service.member.adapter.jpa.MemberEntity;
 
 public interface MemberPort {
