+++ member_service/src/main/java/com/luckytree/member_service/member/application/service/MemberService.java
@@ -1,9 +1,8 @@
 package com.luckytree.member_service.member.application.service;
 
-import com.luckytree.member_service.member.adapter.persistence.MemberEntity;
+import com.luckytree.member_service.member.adapter.jpa.MemberEntity;
 import com.luckytree.member_service.member.application.port.incoming.MemberUseCase;
 import com.luckytree.member_service.member.application.port.outgoing.MemberPort;
-import com.luckytree.member_service.member.domain.Member;
 import com.luckytree.member_service.member.adapter.data.MemberResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
